Refactored XAutocompleteSuggest. #618.
